// enumeratewebapp/src/model/OutsideInfo.ts
export interface OutsideInfo {
  AREA_CODE: string;
  REG: string;
  REGN: string;
  CWT: string;
  CWTN: string;
  AMP: string;
  AMPN: string;
  TAM: string;
  TAMN: string;
  TYPE: string;
  MUN: string;
  VIL: string;
  VILT: string;
  ST_A: string;
  ST_B: string;
  ST_C: string;
  FN_A: string;
  FN_B: string;
  FN_C: string;
  ENU_A: string;
  ENU_B: string;
  ACA_A: string;
  ACA_B: string;
  SUB_A: string;
  SUB_B: string;
  R01: string;
  F1CODE: string;
  OCWT: string;
  OAMP: string;
  OTAM: string;
  OTYPE: string;
  OMUN: string;
  OVIL: string;
  OADD: string;
  KP: string;
  R13_A: string;
  R13_AT: string;
  R13_B: string;
  R13_C: string;
  R17: string;
  R19: string;
  R20: string;
  R21: string;
  R22: string;
  R23: string;
  R24: string;
  R25: string;
  R26: string;
  R27: string;
  AH_CODE: string;
  TempKey: number;
  Temp_AREA_CODE: string;
  Temp_R02: string;
  SDATE_A: string;
  SDATE_B: string;
  SDATE_C: string;
  CDATE_A: string;
  CDATE_B: string;
  CDATE_C: string;
}